v1.2

+++ winenobe.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
+VERSION="1.2-git"
+
+[ "$OSTYPE" != "linux-gnu" ] && echo "WARNING: Operating system not supported ! Use this at your own risk."
 trap 'last_command=$current_command; current_command=$BASH_COMMAND' DEBUG
 
 failed () {
@@ -12,8 +15,18 @@ print_usage () {
 	echo "\"version\" can be either 2012 or 2016."
 	echo
 	echo "Options:"
-	echo "   --help, -h                 Displays this help message and exit."
-	echo "   --skip, -s task            Skips a task."
+	echo "   --help,      -h           Displays this help message and exit."
+	echo "   --version,   -v           Displays the version and exit."
+	echo "   --uninstall, -u comp      Uninstalls the specified component(s)"
+	echo "                             of the specified version."
+	echo
+	echo "   \"comp\" can be one of the following:"
+	echo "   - files: Remove files only."
+	echo "   - reg: Remove registry entries."
+	echo "   - uri: Remove desktop file and URI handler."
+	echo "   - all: Remove all of the above (full removal)."
+	echo
+	echo "   --skip,      -s task      Skips a task."
 	echo
 	echo "   \"task\" can be one of the following:"
 	echo "   - install: skips installation."
@@ -31,6 +44,21 @@ do
 	fi
 done
 
+# checks for wineprefix, if doesn't exist define it anyways.
+if [ -z "$WINEPREFIX" ]
+then
+	WINEPREFIX="$HOME/.wine"
+fi
+
+# determines the program files folder
+W_DRIVE_C="$WINEPREFIX/drive_c"
+if [ -d "$W_DRIVE_C/windows/syswow64" ]
+then
+	W_PROGFILES="$W_DRIVE_C/Program Files (x86)"
+else
+	W_PROGFILES="$W_DRIVE_C/Program Files"
+fi
+
 # check parameters
 case "$1" in
 	"--skip" | "-s")
@@ -45,7 +73,7 @@ case "$1" in
 				echo "Invalid task."
 				echo
 				print_usage
-				exit
+				exit 1
 				;;
 		esac
 		F_VERSION="$3"
@@ -54,16 +82,104 @@ case "$1" in
 		print_usage
 		exit
 		;;
+	"--version" | "-v")
+		echo "Winenobe v$VERSION"
+		exit
+		;;
+	"--uninstall" | "-u")
+		F_VERSION="$3"
+		case "$2" in
+			files)
+				RMFILES=true
+				;;
+			reg)
+				RMREG=true
+				;;
+			uri)
+				RMURI=true
+				;;
+			all)
+				RMFILES=true
+				RMREG=true
+				RMURI=true
+				;;
+			*)
+				echo "Invalid component."
+				echo
+				print_usage
+				exit 1
+				;;
+		esac
+		UNINST=true
+		;;
 	*)
 		F_VERSION="$1"
 		;;
 esac
 
-if [ -z "$F_VERSION" ] && [ "$F_VERSION" != "2012" ] && [ "$F_VERSION" != "2016" ]
+if [ -z "$F_VERSION" ]
 then
 	print_usage
 	exit
 fi
+
+uninstall () {
+	if [ "$F_VERSION" = "2012" ]
+	then
+		INSTALLPATH="$W_DRIVE_C/Finobe/2012"
+		DESKTOPFILE="$HOME/.local/share/applications/finobe-player-2012.desktop"
+		MIMETYPE="x-scheme-handler/finobetw"
+	else
+		INSTALLPATH="$W_PROGFILES/Finobe"
+		DESKTOPFILE="$HOME/.local/share/applications/finobe-player-2016.desktop"
+		MIMETYPE="x-scheme-handler/finobesi"
+	fi
+	MIMEENTRY="$MIMETYPE=$(basename "$DESKTOPFILE")"
+	MIMEAPPS="$HOME/.local/share/applications/mimeapps.list"
+
+	if [ ! -z "$RMFILES" ]
+	then
+		echo "Removing Finobe $F_VERSION..."
+		if [ -d "$INSTALLPATH" ]
+		then
+			rm -rf "$INSTALLPATH" || failed
+		else
+			echo "Finobe $F_VERSION installation not found, skipping."
+		fi
+	fi
+
+	if [ ! -z "$RMURI" ]
+	then
+		echo "Removing desktop file and URI handler..."
+		if [ -f "$DESKTOPFILE" ]
+		then
+			rm "$DESKTOPFILE" || failed
+		else
+			echo "Desktop file not found, skipping."
+		fi
+
+		if grep -Fxq "$MIMEENTRY" "$MIMEAPPS"
+		then
+			sed -i "s#$MIMEENTRY##g" "$MIMEAPPS" || failed
+		else
+			echo "URI handler not found, skipping."
+		fi
+	fi
+
+	if [ ! -z "$RMREG" ]
+	then
+		echo "Deleting registries..."
+		if [ "$F_VERSION" = "2012" ]
+		then
+			wine regedit /D "HKEY_CURRENT_USER\\Software\\FinobeLauncher" || failed
+		else
+			wine regedit /D "HKEY_CURRENT_USER\\Software\\Finobe" || failed
+			wine regedit /D "HKEY_CURRENT_USER\\Software\\Finobe_Penelope" || failed
+		fi
+	fi
+	echo "Uninstallation completed successfully."
+	exit
+}
 
 if [ "$F_VERSION" != "2012" ] && [ "$F_VERSION" != "2016" ]
 then
@@ -73,23 +189,10 @@ then
 	exit 1
 fi
 
-# checks for wineprefix, if doesn't exist define it anyways.
-if [ -z "$WINEPREFIX" ]
-then
-	WINEPREFIX="$HOME/.wine"
-fi
-
-# determines program files folder
-W_DRIVE_C="$WINEPREFIX/drive_c"
-if [ -d "$W_DRIVE_C/windows/syswow64" ]
-then
-	W_PROGFILES="$W_DRIVE_C/Program Files (x86)"
-else
-	W_PROGFILES="$W_DRIVE_C/Program Files"
-fi
+[ ! -z "$UNINST" ] && uninstall
 
 finobe_installed () {
-	echo "Finobe $F_VERSION is already installed, skipping installation."
+	echo "Finobe $F_VERSION is already installed, skipping."
 	echo
 	F_INSTALLED=true
 }
@@ -127,10 +230,12 @@ progressfilt () {
 
 if [ -z "$F_INSTALLED" ] && [ -z "$SKIPINSTALL" ]
 then
+	# hi there, random person. if you're here to check the validity
+	# of those file links below, i can assure you that those files are
+	# the original, unmodified version of them. you can check it for yourself if you want.
 	if [ "$F_VERSION" = "2012" ]
 	then
 		echo "NOTICE: While this script can install the 2012 client, it is not guaranteed to be the lastest version."
-		echo "Also, as of the current wine version (4.15), the 2012 client does not work."
 		read -p "Press Enter to continue or CTRL+C to exit..."
 		echo
 		DLINK="https://files.catbox.moe/ttlil8.bin"
@@ -143,8 +248,12 @@ then
 	echo "The installation process will now begin. If Finobe Studio launches after installing, close it."
 	if [ "$F_VERSION" = "2012" ]
 	then
+		echo
 		echo "WARNING: DO NOT CHANGE THE INSTALLATION DIRECTORY !"
+		echo "Refer to the repo's README for instructions on how to join 2012 games."
 	fi
+	echo
+	echo "Close all other Wine processes before proceeding !"
 	read -p "Press Enter to begin installing."
 	echo "Installing..."
 	wine "install.exe" || failed
@@ -169,8 +278,8 @@ then
 	echo
 fi
 
-if [ -z "$SKIPURI" ]
-then
+[ ! -z "$SKIPURI" ] && exit
+
 if [ "$F_VERSION" = "2016" ]
 then
 	DESKTOPFILE="$HOME/.local/share/applications/finobe-player-2016.desktop"
@@ -192,7 +301,12 @@ then
 else
 	MIMETYPE="x-scheme-handler/finobetw"
 fi
-cat << EOF > "$DESKTOPFILE"
+
+if [ -f "$DESKTOPFILE" ]
+then
+	echo "Desktop file already added, skipping."
+else
+	cat << EOF > "$DESKTOPFILE"
 [Desktop Entry]
 Version=1
 Type=Application
@@ -205,7 +319,8 @@ Actions=
 MimeType=$MIMETYPE
 Categories=Game
 EOF
-echo "Created desktop file."
+	echo "Created desktop file."
+fi
 
 MIMEAPPS="$HOME/.local/share/applications/mimeapps.list"
 if [ ! -f "$MIMEAPPS" ]
@@ -217,10 +332,14 @@ if ! grep -Fxq "[Default Applications]" "$MIMEAPPS"
 then
 	echo "[Default Applications]" >> "$MIMEAPPS" || failed
 fi
-echo >> "$MIMEAPPS" || failed
-echo "$MIMETYPE=$(basename "$DESKTOPFILE")" >> "$MIMEAPPS"
-echo "Added Finobe URI handler."
-echo
-fi
 
-echo "Tasks completed successfully."
+MIMEENTRY="$MIMETYPE=$(basename "$DESKTOPFILE")"
+if grep -Fxq "$MIMEENTRY" "$MIMEAPPS"
+then
+	echo "Finobe URI handler already added, skipping."
+else
+	echo >> "$MIMEAPPS" || failed
+	echo "$MIMEENTRY" >> "$MIMEAPPS"
+	echo "Added Finobe URI handler."
+	echo
+fi
